Refactor and update BA customer service audit implementation: rename file to lowercase, update logic and structure, ensure compliance with framework and ethical standards.
