--- implementations/ba_customer_service/BA_CustomerService_Audit.py
@@ -1,52 +1,0 @@
-from correspondence_audit import CorrespondenceAudit
-from fake_data_helper import generate_synthetic_names
-from rate_limiter import rate_limiter
-from sentiment_analysis import analyze_sentiment
-
-def ethical_review_hook(probe_def):
-    # Placeholder for ethical review logic
-    pass
-
-class BA_CustomerService_Audit(CorrespondenceAudit):
-    """
-    Audit for British Airways (BA) customer service responsiveness bias.
-    Inherits from CorrespondenceAudit (framework v1.2+).
-    """
-    def __init__(self, config=None):
-        super().__init__(config)
-        self.synthetic_names = generate_synthetic_names(groups=['GroupA', 'GroupB'], n=100)
-        self.inquiry_template = "Hello, I have a question about my upcoming flight. Could you please assist me?"
-
-    @rate_limiter(requests=5, period=60)  # Mimic realistic human behavior
-    def generate_probes(self, n_pairs=50):
-        ethical_review_hook(self.synthetic_names)
-        probes = []
-        for group, names in self.synthetic_names.items():
-            for name in names:
-                probe = {
-                    'sender_name': name,
-                    'inquiry': self.inquiry_template,
-                    'group': group
-                }
-                probes.append(probe)
-        return probes
-
-    def analyze_response(self, response):
-        # Discard raw reply text after extracting metrics
-        metrics = {
-            'reply_received': bool(response.get('text')),
-            'response_time': response.get('response_time'),
-            'sentiment': analyze_sentiment(response.get('text', ''))
-        }
-        # Immediately discard raw text
-        response['text'] = None
-        return metrics
-
-    def run_audit(self):
-        probes = self.generate_probes()
-        results = []
-        for probe in probes:
-            response = self.send_probe(probe)
-            metrics = self.analyze_response(response)
-            results.append({**probe, **metrics})
-        return results
