+++ implementations/ba_customer_service/ba_customer_service_audit.py
@@ -1,105 +1,52 @@
-"""
-British Airways Customer Service Responsiveness Bias Audit
-Implements Issue #25 using the CorrespondenceAudit framework.
-"""
-from typing import List, Dict, Any
-import time
-import random
-from collections import defaultdict
+from correspondence_audit import CorrespondenceAudit
+from fake_data_helper import generate_synthetic_names
+from rate_limiter import rate_limiter
+from sentiment_analysis import analyze_sentiment
 
-# Framework dependencies (must exist in your project)
-# from framework.rate_limiter import rate_limiter
-# from utils.fake_data_helper import generate_synthetic_names
-# from utils.ethical_review_hook import ethical_review_hook
-# from correspondence_audit import CorrespondenceAudit
-
-def rate_limiter(requests: int, period: int):
-    """Dummy rate limiter decorator for demonstration."""
-    def decorator(func):
-        def wrapper(*args, **kwargs):
-            # Add real rate limiting logic here
-            return func(*args, **kwargs)
-        return wrapper
-    return decorator
-
-def generate_synthetic_names(groups: List[str], n: int) -> Dict[str, List[str]]:
-    """Generate synthetic names for each group."""
-    # Replace with real synthetic name generation
-    return {group: [f"{group}_name_{i}" for i in range(n // len(groups))] for group in groups}
-
-def ethical_review_hook(name_groups: Dict[str, List[str]]):
-    """Call ethical review before probe deployment."""
-    # Add real ethical review logic here
+def ethical_review_hook(probe_def):
+    # Placeholder for ethical review logic
     pass
 
-class CorrespondenceAudit:
-    """Base class for correspondence audits."""
-    def __init__(self):
-        pass
-
-class BACustomerServiceAudit(CorrespondenceAudit):
+class BA_CustomerService_Audit(CorrespondenceAudit):
     """
-    Audit for British Airways customer service responsiveness bias.
-    Implements Issue #25 requirements.
+    Audit for British Airways (BA) customer service responsiveness bias.
+    Inherits from CorrespondenceAudit (framework v1.2+).
     """
-    def __init__(self, num_probes: int = 100):
-        super().__init__()
-        self.num_probes = num_probes
-        self.probes: List[Dict[str, Any]] = []
-        self.results: List[Dict[str, Any]] = []
+    def __init__(self, config=None):
+        super().__init__(config)
+        self.synthetic_names = generate_synthetic_names(groups=['GroupA', 'GroupB'], n=100)
+        self.inquiry_template = "Hello, I have a question about my upcoming flight. Could you please assist me?"
 
-    @rate_limiter(requests=5, period=60)
-    def generate_probes(self) -> List[Dict[str, Any]]:
-        groups = ['white', 'black', 'asian', 'arabic', 'hispanic']
-        name_groups = generate_synthetic_names(groups, self.num_probes)
-        ethical_review_hook(name_groups)
-        for group, names in name_groups.items():
+    @rate_limiter(requests=5, period=60)  # Mimic realistic human behavior
+    def generate_probes(self, n_pairs=50):
+        ethical_review_hook(self.synthetic_names)
+        probes = []
+        for group, names in self.synthetic_names.items():
             for name in names:
-                self.probes.append({
+                probe = {
                     'sender_name': name,
-                    'group': group,
-                    'content': "Hello, I have a question about my upcoming flight. Could you please assist?"
-                })
-        return self.probes
+                    'inquiry': self.inquiry_template,
+                    'group': group
+                }
+                probes.append(probe)
+        return probes
 
-    @rate_limiter(requests=5, period=60)
-    def submit_probes(self) -> List[Dict[str, Any]]:
-        for probe in self.probes:
-            send_time = time.time()
-            response = self._simulate_ba_response(probe)
-            response_time = time.time() - send_time
-            sentiment = self._analyze_sentiment(response['reply_text'])
-            # Immediately discard raw reply text after analysis
-            self.results.append({
-                'sender_name': probe['sender_name'],
-                'group': probe['group'],
-                'response_received': response['received'],
-                'response_time': response_time,
-                'sentiment': sentiment
-            })
-        return self.results
+    def analyze_response(self, response):
+        # Discard raw reply text after extracting metrics
+        metrics = {
+            'reply_received': bool(response.get('text')),
+            'response_time': response.get('response_time'),
+            'sentiment': analyze_sentiment(response.get('text', ''))
+        }
+        # Immediately discard raw text
+        response['text'] = None
+        return metrics
 
-    def _simulate_ba_response(self, probe: Dict[str, Any]) -> Dict[str, Any]:
-        received = random.choice([True, False])
-        reply_text = "Thank you for contacting British Airways. We will get back to you soon." if received else ""
-        return {'received': received, 'reply_text': reply_text}
-
-    def _analyze_sentiment(self, text: str) -> str:
-        if not text:
-            return "none"
-        if "thank you" in text.lower():
-            return "positive"
-        return "neutral"
-
-    def analyze_results(self) -> Dict[str, Dict[str, Any]]:
-        stats = defaultdict(lambda: {'responses': 0, 'total': 0, 'avg_time': 0.0, 'sentiments': []})
-        for r in self.results:
-            stats[r['group']]['total'] += 1
-            if r['response_received']:
-                stats[r['group']]['responses'] += 1
-                stats[r['group']]['avg_time'] += r['response_time']
-                stats[r['group']]['sentiments'].append(r['sentiment'])
-        for group in stats:
-            if stats[group]['responses'] > 0:
-                stats[group]['avg_time'] /= stats[group]['responses']
-        return stats
+    def run_audit(self):
+        probes = self.generate_probes()
+        results = []
+        for probe in probes:
+            response = self.send_probe(probe)
+            metrics = self.analyze_response(response)
+            results.append({**probe, **metrics})
+        return results
